requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
python-dotenv==1.0.0
langgraph>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
